requirements.txt
Flask
Flask-Login
Werkzeug
python-dotenv
Flask-SocketIO
eventlet
